to-v4: Moving user plugins
